refactor(daemon): PR #173 리뷰 피드백 반영
- 트랜잭션 카운팅 JSON 문자열 패턴 매칭의 fragile 경고 주석 추가
- handle_client 파라미터 과다에 대한 리팩토링 TODO 주석 추가
- active_connections 타입을 usize에서 u32로 통일

+++ crates/proxy_daemon/src/protocol.rs
@@ -82,7 +82,7 @@ pub enum DaemonMessage {
     #[serde(rename = "health_check_result")]
     HealthCheckResult {
         uptime_secs: u64,
-        active_connections: usize,
+        active_connections: u32,
         total_transactions: u64,
     },
     /// 데몬 연결이 끊어졌음을 알리는 메시지
